renamed routes/api/index.js to news.js
